Move files into dirs.
